# backend/openapi.yaml
openapi: 3.1.0
info:
  title: WeTravel API
  description: Python backend with Clerk sessions and Supabase. Authorize with a session
    JWT, never a secret key.
  version: 2.0.0
paths:
...
  /api/journeys:
    get:
      tags:
      - Journeys
      summary: List Journeys
      operationId: list_journeys_api_journeys_get
      security:
      - ClerkSession: []
      parameters:
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_JourneyResponse_'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Too Many Requests
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
    post:
      tags:
      - Journeys
      summary: Create Journey
      operationId: create_journey_api_journeys_post
      security:
      - ClerkSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JourneyInput'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JourneyResponse'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Too Many Requests
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
  /api/journeys/{journey_id}:
    get:
      tags:
      - Journeys
      summary: Get Journey
      operationId: get_journey_api_journeys__journey_id__get
      security:
      - ClerkSession: []
      parameters:
      - name: journey_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 128
          pattern: ^[a-zA-Z0-9_-]+$
          title: Journey Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JourneyResponse'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Too Many Requests
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
    put:
      tags:
      - Journeys
      summary: Update Journey
      description: Replace the full editable trip. Include all stops that should remain.
      operationId: update_journey_api_journeys__journey_id__put
      security:
      - ClerkSession: []
      parameters:
      - name: journey_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 128
          pattern: ^[a-zA-Z0-9_-]+$
          title: Journey Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JourneyInput'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JourneyResponse'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Too Many Requests
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
    delete:
      tags:
      - Journeys
      summary: Delete Journey
      operationId: delete_journey_api_journeys__journey_id__delete
      security:
      - ClerkSession: []
      parameters:
      - name: journey_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 128
          pattern: ^[a-zA-Z0-9_-]+$
          title: Journey Id
      responses:
        '204':
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Too Many Requests
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
  /api/journeys/{journey_id}/publish:
    post:
      tags:
      - Journeys
      summary: Publish Journey
      description: 'Publish visited journal content, excluding private reservation
        details.


        Requires Completed status, a past/current end date, a profile and a visited
        stop.

        Republishing refreshes the same snapshot. These rules are enforced in SQL.'
      operationId: publish_journey_api_journeys__journey_id__publish_post
      security:
      - ClerkSession: []
      parameters:
      - name: journey_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 128
          pattern: ^[a-zA-Z0-9_-]+$
          title: Journey Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishInput'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishResponse'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Too Many Requests
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
...
  /api/media/upload:
    post:
      tags:
      - Media
      summary: Create Upload
      description: Return a direct upload URL/token. JPEG, PNG, WebP only; max 10
        MiB in Storage.
      operationId: create_upload_api_media_upload_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadInput'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Request Entity Too Large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Bad Gateway
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - ClerkSession: []
  /api/media/read:
    post:
      tags:
      - Media
      summary: Read Photo
      description: Signed URL valid for 300 seconds; ownership/publication checked
        by Storage RLS.
      operationId: read_photo_api_media_read_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReadMediaInput'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadMediaResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Request Entity Too Large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Bad Gateway
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - ClerkSession: []
components:
  schemas:
    BookingStatus:
      type: string
      enum:
      - Not booked
      - Booked
      - No booking needed
      - Cancelled
      title: BookingStatus
...
    ErrorResponse:
      properties:
        error:
          type: string
          title: Error
        issues:
          anyOf:
          - items:
              additionalProperties: true
              type: object
            type: array
          - type: 'null'
          title: Issues
      type: object
      required:
      - error
      title: ErrorResponse
...
    JourneyInput:
      properties:
        title:
          type: string
          maxLength: 120
          minLength: 1
          title: Title
        destination:
          type: string
          maxLength: 160
          minLength: 1
          title: Destination
        startDate:
          type: string
          title: Startdate
          examples:
          - '2026-08-01'
        endDate:
          type: string
          title: Enddate
          examples:
          - '2026-08-01'
        status:
          $ref: '#/components/schemas/JourneyStatus'
          default: Planning
        stops:
          items:
            $ref: '#/components/schemas/StopInput'
          type: array
          maxItems: 1000
          title: Stops
      additionalProperties: false
      type: object
      required:
      - title
      - destination
      - startDate
      - endDate
      title: JourneyInput
    JourneyResponse:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        destination:
          type: string
          title: Destination
        startDate:
          type: string
          title: Startdate
        endDate:
          type: string
          title: Enddate
        status:
          $ref: '#/components/schemas/JourneyStatus'
        stops:
          items:
            $ref: '#/components/schemas/StopInput'
          type: array
          title: Stops
        source:
          anyOf:
          - $ref: '#/components/schemas/SourceResponse'
          - type: 'null'
        publishedId:
          anyOf:
          - type: string
          - type: 'null'
          title: Publishedid
      type: object
      required:
      - id
      - title
      - destination
      - startDate
      - endDate
      - status
      - stops
      title: JourneyResponse
    JourneyStatus:
      type: string
      enum:
      - Planning
      - Traveling
      - Completed
      title: JourneyStatus
...
    Page_JourneyResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/JourneyResponse'
          type: array
          title: Items
        offset:
          type: integer
          title: Offset
        limit:
          type: integer
          title: Limit
      type: object
      required:
      - items
      - offset
      - limit
      title: Page[JourneyResponse]
...
    PublishInput:
      properties:
        country:
          type: string
          maxLength: 100
          minLength: 1
          title: Country
        caption:
          type: string
          maxLength: 2000
          title: Caption
          default: ''
      additionalProperties: false
      type: object
      required:
      - country
      title: PublishInput
    PublishResponse:
      properties:
        id:
          type: string
          title: Id
      type: object
      required:
      - id
      title: PublishResponse
...
    ReadMediaInput:
      properties:
        path:
          type: string
          maxLength: 400
          pattern: ^[\w/-]+\.(jpg|png|webp)$
          title: Path
      additionalProperties: false
      type: object
      required:
      - path
      title: ReadMediaInput
    ReadMediaResponse:
      properties:
        signedUrl:
          type: string
          title: Signedurl
      type: object
      required:
      - signedUrl
      title: ReadMediaResponse
...
    SourceResponse:
      properties:
        id:
          type: string
          title: Id
        author:
          anyOf:
          - type: string
          - type: 'null'
          title: Author
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
      type: object
      required:
      - id
      title: SourceResponse
    StopInput:
      properties:
        id:
          type: string
          maxLength: 128
          minLength: 1
          pattern: ^[a-zA-Z0-9_-]+$
          title: Id
        kind:
          $ref: '#/components/schemas/StopKind'
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        day:
          type: integer
          maximum: 365.0
          minimum: 0.0
          title: Day
        endDay:
          anyOf:
          - type: integer
            maximum: 365.0
            minimum: 1.0
          - type: 'null'
          title: Endday
        booking:
          $ref: '#/components/schemas/BookingStatus'
          default: Not booked
        visited:
          type: boolean
          title: Visited
          default: false
        confirmation:
          type: string
          maxLength: 1000
          title: Confirmation
          default: ''
        bookingLink:
          type: string
          title: Bookinglink
          default: ''
        cost:
          type: string
          pattern: ^(?:\d{1,10}(?:\.\d{1,2})?)?$
          title: Cost
          default: ''
        currency:
          type: string
          pattern: ^[A-Z]{3}$
          title: Currency
          default: USD
        cancellationDate:
          type: string
          title: Cancellationdate
          default: ''
        review:
          type: string
          maxLength: 2000
          title: Review
          default: ''
        rating:
          type: integer
          maximum: 5.0
          minimum: 0.0
          title: Rating
          default: 0
        photoUri:
          anyOf:
          - type: string
            maxLength: 400
          - type: 'null'
          title: Photouri
      additionalProperties: false
      type: object
      required:
      - id
      - kind
      - name
      - day
      title: StopInput
    StopKind:
      type: string
      enum:
      - Hotel
      - Restaurant
      - Activity
      title: StopKind
    UploadInput:
      properties:
        extension:
          type: string
          enum:
          - jpg
          - png
          - webp
          title: Extension
      additionalProperties: false
      type: object
      required:
      - extension
      title: UploadInput
    UploadResponse:
      properties:
        path:
          type: string
          title: Path
        signedUrl:
          type: string
          title: Signedurl
        token:
          type: string
          title: Token
      type: object
      required:
      - path
      - signedUrl
      - token
      title: UploadResponse
...
  securitySchemes:
    ClerkSession:
      type: http
      description: Clerk session JWT from getToken(); never your secret key.
      scheme: bearer